backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
pyjwt==2.9.0
bcrypt==4.2.0
boto3==1.35.19
python-multipart==0.0.9
